Warning compilation
Former-commit-id: 15806ca84dc512ee4c0e8b92d8500e392aedb685

+++ applications/programs/image_vectorize/image_vectorize_main.cpp
@@ -111,7 +111,7 @@ public:
                     if (apply_mask)
                     {
                         mask.generate_mask(ZSIZE(img()),YSIZE(img()),XSIZE(img()));
-                        vectorSize=mask.get_binary_mask().sum();
+                        vectorSize=(int)mask.get_binary_mask().sum();
                     }
                     else
                         vectorSize=MULTIDIM_SIZE(img());
